refactor(app): move AppState construction/bootstrap out of mod.rs
CODE-4: relocate new_with_drivers_and_settings, the cfg-gated driver
registration, audit-settings bootstrap and auth-secret hydration into
app_state/bootstrap.rs, leaving accessors and runtime methods in mod.rs.
Feature gates preserved verbatim. Pure code motion; no public API change.
